requirements.txt
easydict
ftfy
git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup.git
git+https://github.com/openai/CLIP.git
mlflow
numpy==1.23.3
open3d
opencv-python
open_clip_torch
graphviz
plyfile
pytorch-lightning==2.1.1
scipy
setuptools
tensorflow==2.12.0
torch-cluster==1.6.3
torch-geometric==2.5.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
tqdm
transformers>=4.31.0
trimesh